usb: move all control-related stuff to `mod control`.
